+++ test/Utilities/sets.jl
@@ -1,3 +1,5 @@
+using SparseArrays
+
 @testset "Constant" begin
     @test MOI.constant(MOI.EqualTo(3)) == 3
     @test MOI.constant(MOI.GreaterThan(6)) == 6
@@ -91,4 +93,94 @@ end
     s = MOI.LogDetConeTriangle(4)
     err = ErrorException("Dual of $s is not implemented.")
     @test_throws err MOI.dual_set(MOI.LogDetConeTriangle(4))
+end
+
+@testset "Set dot" begin
+    vec = zeros(6)
+    @test MOIU.set_dot(vec, vec, MOI.SecondOrderCone(6)) == 0
+    @test MOIU.set_dot(vec, vec, MOI.PositiveSemidefiniteConeTriangle(3)) == 0
+    @test MOIU.set_dot(vec, vec, MOI.PositiveSemidefiniteConeSquare(3)) == 0
+    vec[5] = 1
+    @test MOIU.set_dot(vec, vec, MOI.SecondOrderCone(6)) == 1
+    @test MOIU.set_dot(vec, vec, MOI.PositiveSemidefiniteConeTriangle(3)) == 2
+    @test MOIU.set_dot(vec, vec, MOI.PositiveSemidefiniteConeSquare(3)) == 1
+    vec[5] = 0
+    vec[3] = 1
+    @test MOIU.set_dot(vec, vec, MOI.SecondOrderCone(6)) == 1
+    @test MOIU.set_dot(vec, vec, MOI.PositiveSemidefiniteConeTriangle(3)) == 1
+    @test MOIU.set_dot(vec, vec, MOI.PositiveSemidefiniteConeSquare(3)) == 1
+
+    vec = zeros(7)
+    @test MOIU.set_dot(vec, vec, MOI.RootDetConeTriangle(3)) == 0
+    vec[5] = 1
+    @test MOIU.set_dot(vec, vec, MOI.RootDetConeTriangle(3)) == 2
+    vec = zeros(8)
+    @test MOIU. set_dot(vec, vec, MOI.LogDetConeTriangle(3)) == 0
+    vec[5] = 1
+    @test MOIU.set_dot(vec, vec, MOI.LogDetConeTriangle(3)) == 1
+    
+    sp_vec = spzeros(6)
+    @test MOIU.set_dot(sp_vec, sp_vec, MOI.SecondOrderCone(6)) == 0
+    @test MOIU.set_dot(sp_vec, sp_vec, MOI.PositiveSemidefiniteConeTriangle(3)) == 0
+    @test MOIU.set_dot(sp_vec, sp_vec, MOI.PositiveSemidefiniteConeSquare(3)) == 0
+    sp_vec[5] = 1
+    @test MOIU.set_dot(sp_vec, sp_vec, MOI.SecondOrderCone(6)) == 1
+    @test MOIU.set_dot(sp_vec, sp_vec, MOI.PositiveSemidefiniteConeTriangle(3)) == 2
+    @test MOIU.set_dot(sp_vec, sp_vec, MOI.PositiveSemidefiniteConeSquare(3)) == 1
+    sp_vec[5] = 0
+    sp_vec[3] = 1
+    @test MOIU.set_dot(sp_vec, sp_vec, MOI.SecondOrderCone(6)) == 1
+    @test MOIU.set_dot(sp_vec, sp_vec, MOI.PositiveSemidefiniteConeTriangle(3)) == 1
+    @test MOIU.set_dot(sp_vec, sp_vec, MOI.PositiveSemidefiniteConeSquare(3)) == 1
+
+    sp_vec = spzeros(7)
+    @test MOIU.set_dot(sp_vec, sp_vec, MOI.RootDetConeTriangle(3)) == 0
+    sp_vec[5] = 1
+    @test MOIU.set_dot(sp_vec, sp_vec, MOI.RootDetConeTriangle(3)) == 2
+    sp_vec = spzeros(8)
+    @test MOIU. set_dot(sp_vec, sp_vec, MOI.LogDetConeTriangle(3)) == 0
+    sp_vec[5] = 1
+    @test MOIU.set_dot(sp_vec, sp_vec, MOI.LogDetConeTriangle(3)) == 1
+end
+
+@testset "dot coefficients" begin
+    vec = zeros(6)
+    @test MOIU.dot_coefficients(vec, MOI.SecondOrderCone(6)) == vec
+    @test MOIU.dot_coefficients(vec, MOI.PositiveSemidefiniteConeTriangle(3)) == vec
+    vec[5] = 1
+    @test MOIU.dot_coefficients(vec, MOI.SecondOrderCone(6)) == vec
+    @test MOIU.dot_coefficients(vec, MOI.PositiveSemidefiniteConeTriangle(3)) == vec./2
+    vec[5] = 0
+    vec[3] = 1
+    @test MOIU.dot_coefficients(vec, MOI.SecondOrderCone(6)) == vec
+    @test MOIU.dot_coefficients(vec, MOI.PositiveSemidefiniteConeTriangle(3)) == vec
+
+    vec = zeros(7)
+    @test MOIU.dot_coefficients(vec, MOI.RootDetConeTriangle(3)) == vec
+    vec[5] = 1
+    @test MOIU.dot_coefficients(vec, MOI.RootDetConeTriangle(3)) == vec./2
+    vec = zeros(8)
+    @test MOIU. dot_coefficients(vec, MOI.LogDetConeTriangle(3)) == vec
+    vec[5] = 1
+    @test MOIU.dot_coefficients(vec, MOI.LogDetConeTriangle(3)) == vec
+
+    sp_vec = spzeros(6)
+    @test MOIU.dot_coefficients(sp_vec, MOI.SecondOrderCone(6)) == sp_vec
+    @test MOIU.dot_coefficients(sp_vec, MOI.PositiveSemidefiniteConeTriangle(3)) == sp_vec
+    sp_vec[5] = 1
+    @test MOIU.dot_coefficients(sp_vec, MOI.SecondOrderCone(6)) == sp_vec
+    @test MOIU.dot_coefficients(sp_vec, MOI.PositiveSemidefiniteConeTriangle(3)) == sp_vec./2
+    sp_vec[5] = 0
+    sp_vec[3] = 1
+    @test MOIU.dot_coefficients(sp_vec, MOI.SecondOrderCone(6)) == sp_vec
+    @test MOIU.dot_coefficients(sp_vec, MOI.PositiveSemidefiniteConeTriangle(3)) == sp_vec
+
+    sp_vec = spzeros(7)
+    @test MOIU.dot_coefficients(sp_vec, MOI.RootDetConeTriangle(3)) == sp_vec
+    sp_vec[5] = 1
+    @test MOIU.dot_coefficients(sp_vec, MOI.RootDetConeTriangle(3)) == sp_vec./2
+    sp_vec = spzeros(8)
+    @test MOIU. dot_coefficients(sp_vec, MOI.LogDetConeTriangle(3)) == sp_vec
+    sp_vec[5] = 1
+    @test MOIU.dot_coefficients(sp_vec, MOI.LogDetConeTriangle(3)) == sp_vec
 end